+++ test/src/Exception/DumpModNotLoadedExceptionTest.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+namespace FactorioItemBrowserTest\Export\Exception;
+
+use Exception;
+use FactorioItemBrowser\Export\Exception\DumpModNotLoadedException;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * The PHPUnit test of the DumpModNotLoadedException class.
+ *
+ * @author BluePsyduck <bluepsyduck@gmx.com>
+ * @license http://opensource.org/licenses/GPL-3.0 GPL v3
+ * @coversDefaultClass \FactorioItemBrowser\Export\Exception\DumpModNotLoadedException
+ */
+class DumpModNotLoadedExceptionTest extends TestCase
+{
+    /**
+     * Tests the constructing.
+     * @covers ::__construct
+     */
+    public function testConstruct(): void
+    {
+        $expectedMessage = 'The dump mod was not loaded as last mod. This is an indication that at least one mod '
+            . 'is not compatible with the used version of Factorio.';
+
+        /* @var Exception&MockObject $previous */
+        $previous = $this->createMock(Exception::class);
+
+        $exception = new DumpModNotLoadedException($previous);
+
+        $this->assertSame($expectedMessage, $exception->getMessage());
+        $this->assertSame($previous, $exception->getPrevious());
+    }
+}